pyproject.toml
[tool.poetry]
name = "etl"
version = "0.1.0"
description = ""
authors = ["RikiSot <rkgomezz@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
ipykernel = "^6.13.0"
pandas = "^1.4.2"
mysql-connector-python = "^8.0.29"
python-dotenv = "^0.20.0"
SQLAlchemy = "^1.4.36"
pymongo = "^4.1.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
ipykernel = "^6.13.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
